Add AMD64 SSE slice operations
This adds assembler functions for common single and double precision operations. The assembler functions usually are more than twice as fast as pure Go operations. The speedups listed in Xslice.go is on an array with 10000 elements.

If you think we should move on with this, I will add tests and license, and whatever you think also may be missing.